[update-integration-test] skip some test case
